refactor: module file camel case changed to kebab case
