+++ cli/module-generator.js
@@ -15,6 +15,10 @@ if (!moduleName) {
 const nameCamel = moduleName.charAt(0).toLowerCase() + moduleName.slice(1);
 const namePascal = moduleName.charAt(0).toUpperCase() + moduleName.slice(1);
 
+const nameKebab = moduleName
+    .replace(/([a-z0-9])([A-Z])/g, "$1-$2")
+    .toLowerCase();
+
 const modulePath = path.join(process.cwd(), "src/modules", nameCamel);
 const schemaPath = path.join(process.cwd(), "src/lib/validation", nameCamel);
 
@@ -25,7 +29,7 @@ const diContainerPath = path.join(
 
 const files = ["handler.ts", "service.ts", "route.ts", "index.ts"];
 
-if (fs.existsSync(nameCamel)) {
+if (fs.existsSync(modulePath)) {
     console.error(`❌ Module "${nameCamel}" already exists!`);
     process.exit(1);
 }
@@ -35,26 +39,28 @@ console.log(`📁 Created folder: ${modulePath}`);
 
 files.forEach((file) => {
     const fileType = file.replace(".ts", "");
-    const filePath = path.join(modulePath, `${nameCamel}.${file}`); // route.ts -> moduleName.route.ts
+    const filePath = path.join(modulePath, `${nameKebab}.${file}`); // route.ts -> moduleName.route.ts
 
     let content = `// ${file} for ${namePascal}\n`;
 
     // if templates object contains fileType as a key
     if (templates[fileType]) {
-        content = templates[fileType](namePascal, nameCamel);
+        content = templates[fileType](namePascal, nameCamel, nameKebab);
     }
 
     fs.writeFileSync(filePath, content);
     console.log(`📄 Created file: ${filePath}`);
 });
 
+// create schema folder
 fs.mkdirSync(schemaPath, { recursive: true });
 console.log(`📁 Created schema folder: ${schemaPath}`);
-
-const schemaFilePath = path.join(schemaPath, `${nameCamel}.schema.ts`);
-fs.writeFileSync(schemaFilePath, `// Schema for ${namePascal}\n`);
+// create schema file
+const schemaFilePath = path.join(schemaPath, `${nameKebab}.schema.ts`);
+fs.writeFileSync(schemaFilePath, `// Schema for ${nameKebab}\n`);
 console.log("✅ Schema created successfully!");
 
+// read content from di-container.ts
 let diContent = fs.readFileSync(diContainerPath, "utf8");
 
 // Insert imports if not already there
@@ -62,8 +68,8 @@ if (!diContent.includes(`${namePascal}Service`)) {
     const importMarker = 'import { EnvConfig } from "./env.type.js";';
 
     const newImports = `
-import { ${namePascal}Service } from "@/modules/${nameCamel}/${nameCamel}.service.js";
-import { ${namePascal}Handler } from "@/modules/${nameCamel}/${nameCamel}.handler.js";`;
+import { ${namePascal}Service } from "@/modules/${nameCamel}/${nameKebab}.service.js";
+import { ${namePascal}Handler } from "@/modules/${nameCamel}/${nameKebab}.handler.js";`;
 
     diContent = diContent.replace(importMarker, `${importMarker}${newImports}`);
 }
